Archive subsystem: Separate from Spec
