+++ fsspec/implementations/libarchive.py
@@ -14,7 +14,8 @@ from ctypes import (
 import libarchive
 import libarchive.ffi as ffi
 
-from fsspec import AbstractArchiveFileSystem, open_files
+from fsspec import open_files
+from fsspec.archive import AbstractArchiveFileSystem
 from fsspec.implementations.memory import MemoryFile
 from fsspec.utils import DEFAULT_BLOCK_SIZE
